+++ CHANGES.md
@@ -1,4 +1,7 @@
 
+- Fix `Bytes.Slice.compare`. The last byte of equal length slices was
+  not compared (#14). Thanks to Anil Madhavapeddy for the report.
+
 - Fix `Bytes.Reader.of_slice` when the given slice does not start at 0.
   Thanks to Thomas Gazagnaire for report (#13).